Network.Solana.Core.Account

Documentation

newtype Lamport

Lamport is the smallest unit of SOL (1 SOL = 1 billion lamports).

Constructors

Lamport

Fields

unLamport :: Word64

Account

data Account

Constructors

Account

Fields

pubkey :: SolanaPublicKey

account :: AccountInfo

Account Info

data AccountInfo

Constructors

AccountInfo

Fields

lamports :: Lamport

Amount of lamports in the account

dataField :: AccountData

owner :: SolanaPublicKey

executable :: Bool

A boolean flag that indicates whether this account contains a loaded program.

Account Data

data AccountData

A byte array that stores arbitrary data for an account.

Constructors

AccountDataBinary

Fields

accData :: ByteString

AccountDataJSON

Fields

accDataObj :: String