Network.Solana.Core.Instruction

Documentation

data Instruction

mkInstruction :: Binary a => SolanaPublicKey -> [AccountMeta] -> a -> Instruction

iProgramId :: Instruction -> SolanaPublicKey

iAccounts :: Instruction -> [AccountMeta]

List of metadata describing accounts that should be passed to the program.

iData :: Instruction -> InstructionData

data AccountMeta

Each account required by an instruction must be provided as an AccountMeta that contains:

Constructors

AccountMeta

Fields

accountPubKey :: SolanaPublicKey

Account's address

isSigner :: Bool

isWritable :: Bool

newtype InstructionData

Constructors

InstructionData

Fields

instrData :: ByteString

compileInstruction :: [SolanaPublicKey] -> Instruction -> Either CompileException CompiledInstruction

data CompiledInstruction

The structure of a Compiled Instructuin.

data CompileException