Network.Solana.RPC.HTTP.Account

getAccountInfo

getAccountInfo' :: JsonRpc m => SolanaPublicKey -> m (RPCResponse (Maybe AccountInfo))

Fetches all available data for the given account address. Returns 'RpcResponse (Maybe AccountInfo)', where Nothing indicates the account does not exist.

getAccountInfo :: JsonRpc m => SolanaPublicKey -> m (Maybe AccountInfo)

Fetches all available data for the given account address. Returns Nothing if the account does not exist.

getBalance

getBalance' :: JsonRpc m => SolanaPublicKey -> m (RPCResponse Lamport)

Returns the balance, in lamports, of the specified account. Returns 'RpcResponse Lamport'.

getBalance :: JsonRpc m => SolanaPublicKey -> m Lamport

Returns the balance, in lamports, of the specified account.

getMultipleAccounts

getMultipleAccounts' :: JsonRpc m => [SolanaPublicKey] -> m (RPCResponse [Maybe AccountInfo])

Returns account information for a list of addresses. For any missing account, the result contains Nothing in its place.

getMultipleAccounts :: JsonRpc m => [SolanaPublicKey] -> m [Maybe AccountInfo]

Returns account information for a list of addresses. For any missing account, the result contains Nothing in its place.

getProgramAccounts

getProgramAccounts :: JsonRpc m => SolanaPublicKey -> m [Account]

Returns all accounts owned by the specified program address.

getLargestAccounts

data AddressAndLamports

Contains the address and value of an account.

Constructors

AddressAndLamports

Fields

address :: SolanaPublicKey

Account address

lamports :: Lamport

Number of lamports in the account

getLargestAccounts' :: JsonRpc m => m (RPCResponse [AddressAndLamports])

Returns up to the 20 accounts with the highest balances in lamports. Results may be cached for up to two hours.

getLargestAccounts :: JsonRpc m => m [(SolanaPublicKey, Lamport)]

Returns up to the 20 accounts with the highest balances in lamports. Results may be cached for up to two hours.

getSignaturesForAddress

getSignaturesForAddress :: JsonRpc m => SolanaPublicKey -> m [TransactionSignatureInformation]

Returns confirmed transaction signatures involving the given address, in reverse chronological order.

data TransactionSignatureInformation

Metadata for a confirmed transaction signature involving a given address.

Constructors

TransactionSignatureInformation

Fields

signature :: SolanaSignature

Transaction signature

slotTxSig :: Slot

The slot that contains the block with the transaction

err :: Maybe String

Error if transaction failed, Nothing if transaction succeeded.

memo :: Maybe String

Memo associated with the transaction, Nothing if no memo is present

blockTime :: Maybe Int64

Estimated production time, as Unix timestamp (seconds since the Unix epoch) of when transaction was processed. Nothing if not available.

confirmationStatus :: Maybe String

The transaction's cluster confirmation status;

getSignatureStatuses

getSignatureStatuses' :: JsonRpc m => [SolanaSignature] -> ConfigurationObject -> m (RPCResponse [Maybe TransactionSignatureStatus])

Returns the confirmation status and slot info for the specified transaction signatures. Each signature should be the transaction’s first signature (used as a unique identifier).

getSignatureStatuses :: JsonRpc m => [SolanaSignature] -> m [Maybe TransactionSignatureStatus]

Returns the confirmation status and slot info for the specified transaction signatures. Each signature should be the transaction’s first signature (used as a unique identifier).

data TransactionSignatureStatus

Status metadata for a given transaction signature.

Constructors

TransactionSignatureStatus

Fields

slotTxStatus :: Slot

The slot the transaction was processed

confirmationsTxStatus :: Maybe Int

Number of blocks since signature confirmation, Nothing if rooted, as well as finalized by a supermajority of the cluster.

errTxStatus :: Maybe String

Error if transaction failed, Nothing if transaction succeeded.

confirmationStatusTxStatus :: Maybe String

The transaction's cluster confirmation status.