Network.Solana.RPC.HTTP.Block

getBlock

getBlock :: JsonRpc m => Slot -> m BlockInfo

Returns identity and transaction information about a confirmed block in the ledger.

data BlockInfo

Contains identity and transaction information for a block.

Constructors

BlockInfo

Fields

blockhashBI :: BlockHash

The blockhash of this block.

previousBlockhashBI :: BlockHash

The blockhash of this block's parent. If the parent is not available due to ledger cleanup, this will return "11111111111111111111111111111111".

parentSlotBI :: Slot

The slot index of this block's parent.

transactionsBI :: [TransactionWithMeta]

List of transactions included in the block, each with optional metadata.

blockTimeBI :: Maybe Int64

Estimated production time, as a Unix timestamp (in seconds). Nothing if not available.

blockHeightBI :: Maybe BlockHeight

The number of blocks beneath this block.

data TransactionWithMeta

A transaction and its optional metadata as included in a block.

Constructors

TransactionWithMeta

Fields

meta :: Maybe Object

Optional metadata for the transaction.

transaction :: Transaction

The transaction details.

getBlockCommitment

getBlockCommitment :: JsonRpc m => Slot -> m BlockCommitment

Get the block commitment based on the given block number Slot. Returns a BlockCommitment for the specified block.

data BlockCommitment

Commitment information for a block.

Constructors

BlockCommitment

Fields

commitmentList :: Maybe [Word64]

Array logging the amount of cluster stake in lamports that has voted on the block at each depth from 0 to MAX_LOCKOUT_HISTORY

totalStake :: Integer

Total active stake, in lamports, for the current epoch.

getBlockHeight

getBlockHeight :: JsonRpc m => m BlockHeight

Returns the current block height of the node.

getBlockProduction

getBlockProduction' :: JsonRpc m => m (RPCResponse BlockProduction)

Returns recent block production information from the current or previous epoch. Returns 'RpcResponse BlockProduction' with the production statistics.

getBlockProduction :: JsonRpc m => m BlockProduction

Returns recent block production information from the current or previous epoch.

type ValidatorData = [Word64]

Range of slots considered.

type ByIdentity = Map SolanaPublicKey ValidatorData

Mapping from validator identity to number of blocks produced.

data BlockProductionRange

Range of slots considered for block production.

Constructors

BlockProductionRange

Fields

firstSlot :: Slot

First slot in the range.

lastSlot :: Slot

Last slot in the range.

data BlockProduction

Block production statistics mapped by validator identity.

Constructors

BlockProduction

Fields

ByIdentity

Mapping from validator identity to number of blocks produced.

BlockProductionRange

Range of slots considered.

getBlocks

getBlocks :: JsonRpc m => Slot -> Maybe Slot -> m [Int]

Returns a list of confirmed blocks between two slots (inclusive). The maximum range allowed is 500,000 slots.

getBlocksWithLimit

getBlocksWithLimit :: JsonRpc m => Slot -> Int -> m [Int]

Returns a list of confirmed blocks starting from the given slot, up to the specified limit.

getBlockTime

getBlockTime :: JsonRpc m => Slot -> m Int

Returns the estimated production time of a block (as Unix timestamp). This is based on stake-weighted votes and validator-reported timestamps.

getLatestBlockhash

getLatestBlockhash' :: JsonRpc m => m (RPCResponse LatestBlockHash)

Get the hash and height of the latest block. Returns 'RpcResponse LatestBlockHash' containing the block information.

getLatestBlockhash :: JsonRpc m => m LatestBlockHash

Get the hash and height of the latest block.

getTheLatestBlockhash :: JsonRpc m => m BlockHash

Get only the blockhash of the latest block.

data LatestBlockHash

Contains the hash and the height of the latest block.

Constructors

LatestBlockHash

Fields

blockhash :: BlockHash

Block hash as a base-58 encoded string.

lastValidBlockHeight :: BlockHeight

Last block height at which the blockhash is still considered valid.

isBlockhashValid

isBlockhashValid' :: JsonRpc m => BlockHash -> m (RPCResponse Bool)

Returns whether a blockhash is still valid.

isBlockhashValid :: JsonRpc m => BlockHash -> m Bool

Returns whether a blockhash is still valid.