Network.Solana.RPC.HTTP.Chain

getClusterNodes

getClusterNodes :: JsonRpc m => m [ClusterNodes]

Returns information about all the nodes currently participating in the cluster.

data ClusterNodes

Contains information about all the nodes participating in the cluster.

Constructors

ClusterNodes

Fields

pubkey :: SolanaPublicKey

The node's public key.

gossip :: Maybe String

Gossip network address for the node.

tpu :: Maybe String

TPU (Transaction Processing Unit) network address.

rpc :: Maybe String

JSON-RPC address, if the RPC service is enabled.

version :: Maybe String

Software version of the node, if available.

featureSet :: Maybe Word32

Unique identifier of the node's feature set.

sharedVersion :: Maybe Word16

Shred version used by the node.

getHealth

getHealth :: JsonRpc m => m String

Checks if the node is healthy.

A healthy node is within a certain number of slots (defined by HEALTH_CHECK_SLOT_DISTANCE) of the latest confirmed cluster slot. Returns "ok" if healthy; otherwise, an error is returned.

getHighestSnapshotSlot

getHighestSnapshotSlot :: JsonRpc m => m HighestSnapshotSlot

Returns the highest full and incremental snapshot slots that the node has generated.

This includes the highest full snapshot slot and, if available, the corresponding highest incremental snapshot slot.

data HighestSnapshotSlot

Contains the highest slot information that the node has snapshots for.

Constructors

HighestSnapshotSlot

Fields

full :: Word64

The highest full snapshot slot.

incremental :: Maybe Word64

The highest incremental snapshot slot, if available.

getIdentity

getIdentity' :: JsonRpc m => m NodeIdentity

Returns the full NodeIdentity object for the current node.

getIdentity :: JsonRpc m => m SolanaPublicKey

Returns the identity public key of the current node.

newtype NodeIdentity

Identity information for the current node.

Constructors

NodeIdentity

Fields

identity :: SolanaPublicKey

The node's identity public key.

getRecentPerformanceSamples

getRecentPerformanceSamples :: JsonRpc m => Maybe Int -> m [PerformanceSample]

Returns a list of recent PerformanceSample entries, in reverse slot order.

Performance samples are collected every 60 seconds and include metrics such as the number of transactions, slots, and non-vote transactions within each sample window. Optionally takes the number of samples to return (must be ≤ 720).

data PerformanceSample

Performance metrics sampled periodically from the node.

Constructors

PerformanceSample

Fields

slot :: Slot

Slot in which the sample was taken.

numTransactions :: Word64

Number of transactions during the sample period.

numSlots :: Word64

Number of slots completed during the sample.

samplePeriodSecs :: Word16

Duration of the sampling window, in seconds.

numNonVoteTransactions :: Word64

Number of non-vote transactions during the sample.

getRecentPrioritizationFees

getRecentPrioritizationFees :: JsonRpc m => Maybe [SolanaPublicKey] -> m [PrioritizationFee]

Returns recent prioritization fees observed in recent blocks.

Optionally takes a list of up to 128 account addresses. If provided, the response estimates the prioritization fee for landing a transaction that locks all listed accounts as writable.

data PrioritizationFee

A prioritization fee sample from a recent block.

Constructors

PrioritizationFee

Fields

pfSlot :: Slot

Slot in which the fee was observed.

prioritizationFee :: Word64

Per-compute-unit fee (in micro-lamports).

getVersion

getVersion :: JsonRpc m => m SolanaVersion

Returns the Solana software version and feature set currently running on the node.

data SolanaVersion

Contains information about the software version and feature set identifier of the node.

Constructors

SolanaVersion

Fields

solana_core :: String

Software version of 'solana-core'.

feature_set :: Word32

Unique identifier of the software's feature set.