Network.Solana.RPC.HTTP.Ledger

getEpochInfo

getEpochInfo :: JsonRpc m => m EpochInfo

Returns information about the current epoch.

This includes the current slot, block height, epoch index, and transaction count since genesis.

type Epoch = Word64

Solana epoch index.

data EpochInfo

Contains information about the current epoch.

Constructors

EpochInfo

Fields

absoluteSlot :: Slot

The current absolute slot.

blockHeight :: Word64

The current block height.

epoch :: Epoch

The current epoch index.

slotIndex :: Word64

Slot index relative to the start of the epoch.

slotsInEpoch :: Word64

Total number of slots in the epoch.

transactionCount :: Maybe Word64

Total number of successful transactions since genesis.

getEpochSchedule

getEpochSchedule :: JsonRpc m => m EpochSchedule

Returns the epoch scheduling parameters configured in the cluster's genesis configuration.

data EpochSchedule

Contains epoch schedule information as defined in the cluster's genesis config.

Constructors

EpochSchedule

Fields

slotsPerEpoch :: Word64

Maximum number of slots per epoch.

leaderScheduleSlotOffset :: Word64

Number of slots before epoch start to compute leader schedule.

warmup :: Bool

Whether epochs start small and grow (warmup).

firstNormalEpoch :: Word64

First normal-length epoch.(log2(slotsPerEpoch) - log2(MINIMUM_SLOTS_PER_EPOCH))

firstNormalSlot :: Word64

First slot of the first normal-length epoch. ( MINIMUM_SLOTS_PER_EPOCH * (2.pow(firstNormalEpoch) - 1))

getFirstAvailableBlock

getFirstAvailableBlock :: JsonRpc m => m Slot

Returns the slot of the lowest confirmed block that has not been purged from the ledger.

getGenesisHash

getGenesisHash :: JsonRpc m => m BlockHash

Returns the block hash of the genesis block.

getLeaderSchedule

getLeaderSchedule :: JsonRpc m => Maybe Slot -> m (Maybe LeaderSchedule)

Returns the leader schedule mapping validator identities to their leader slots.

Fetches the leader schedule for the epoch that contains the specified slot. If no slot is specified, the schedule for the current epoch is returned.

type LeaderSchedule = Map SolanaPublicKey [Word64]

Map from validator identity pubkeys to arrays of slot indices indicating when each validator is expected to produce a block.

getMaxRetransmitSlot

getMaxRetransmitSlot :: JsonRpc m => m Slot

Returns the highest slot observed by the node from the retransmit stage.

getMaxShredInsertSlot

getMaxShredInsertSlot :: JsonRpc m => m Slot

Returns the highest slot observed by the node after shred insertion.

getSlot

getSlot :: JsonRpc m => m Slot

Returns the current slot, based on the default or specified commitment level.

getSlotLeader

getSlotLeader :: JsonRpc m => m SolanaPublicKey

Returns the identity public key of the current slot leader.

getSlotLeaders

getSlotLeaders :: JsonRpc m => Slot -> Int -> m [SolanaPublicKey]

Returns the list of slot leaders for a given slot range.

Takes a starting Slot and a limit (between 1 and 5,000), and returns the validator identities that will produce blocks for the given slots.

getTransactionCount

getTransactionCount :: JsonRpc m => m Slot

Returns the total number of transactions recorded in the ledger.

minimumLedgerSlot

minimumLedgerSlot :: JsonRpc m => m Slot

Returns the lowest slot that the node has information about in its ledger.